requirements.txt
# AI Video Editor - Dependencies (Fixed NumPy Compatibility)
# Core Flask framework
Flask==3.0.0
Flask-CORS==4.0.0
werkzeug==3.0.1

# AI & Machine Learning - Fixed compatibility
openai-whisper==20231117
torch
torchaudio==2.1.0
transformers==4.35.2
accelerate==0.24.1

# Text-to-Speech
edge-tts==6.1.12
gtts==2.5.1

# Video/Audio Processing with NumPy 1.x compatibility
ffmpeg-python==0.2.0
moviepy==1.0.3
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1

# Image Processing - Updated for compatibility
opencv-python==4.8.1.78
Pillow==10.1.0

# Utilities - FIXED NumPy version for compatibility
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
pathlib

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0

# Multiple TTS Providers
openai==1.3.0
aiohttp==3.9.1

# Optional Premium TTS Providers (install if you have API keys)
# google-cloud-texttospeech==2.14.2
# azure-cognitiveservices-speech==1.34.0

# Web scraping (for ElevenLabs)
python-multipart==0.0.6

# For Google Colab tunneling
pyngrok==7.1.6 